requirements.txt
# for heroku deployment
dj-database-url==0.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.5
whitenoise==5.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
coreapi==2.3.3
